@@ -20,39 +20,36 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^8.1",
+        "php": ">=8.3",
         "dontdrinkandroot/common": "^0.2",
-        "dontdrinkandroot/path": "^0.1.0",
-        "league/commonmark": "^2.1",
-        "sensio/framework-extra-bundle": "^6.2",
-        "symfony/asset": "^5.4",
-        "symfony/config": "^5.4",
-        "symfony/finder": "^5.4",
-        "symfony/form": "^5.4",
-        "symfony/framework-bundle": "^5.4",
-        "symfony/mime": "^5.4",
-        "symfony/templating": "^5.4",
-        "symfony/translation": "^5.4",
-        "symfony/twig-bridge": "^5.4",
-        "symfony/twig-bundle": "^5.4",
-        "symfony/validator": "^5.4",
-        "symfony/yaml": "^5.4",
+        "dontdrinkandroot/path": "^0.2",
+        "league/commonmark": "^2.4",
+        "symfony/asset": "^6",
+        "symfony/config": "^6",
+        "symfony/finder": "^6",
+        "symfony/form": "^6",
+        "symfony/framework-bundle": "^6",
+        "symfony/mime": "^6",
+        "symfony/templating": "^6",
+        "symfony/translation": "^6",
+        "symfony/twig-bridge": "^6",
+        "symfony/twig-bundle": "^6",
+        "symfony/validator": "^6",
+        "symfony/yaml": "^6",
         "symplify/git-wrapper": "^10.0"
     },
     "require-dev": {
         "elasticsearch/elasticsearch": "^7.0",
-        "guzzlehttp/ringphp": "^1.1.1",
-        "liip/test-fixtures-bundle": "^1.0",
-        "phpunit/phpunit": "^9.5.11",
-        "symfony/browser-kit": "^5.4",
-        "symfony/console": "^5.4",
-        "symfony/css-selector": "^5.4",
-        "symfony/phpunit-bridge": "^5.4",
-        "symfony/security-bundle": "^5.4"
+        "liip/test-fixtures-bundle": "^2.7",
+        "phpunit/phpunit": "^9.6",
+        "symfony/browser-kit": "^6",
+        "symfony/console": "^6",
+        "symfony/css-selector": "^6",
+        "symfony/phpunit-bridge": "^6",
+        "symfony/security-bundle": "^6"
     },
     "suggest": {
-        "elasticsearch/elasticsearch": "Include this if you want the wiki to be searchable, Supported Version: ^7.0",
-        "braincrafted/bootstrap-bundle": "Style forms in a bootstrap compatible style"
+        "elasticsearch/elasticsearch": "Include this if you want the wiki to be searchable, Supported Version: ^7.0"
     },
     "autoload": {
         "psr-4": {
@@ -66,11 +63,11 @@
     },
     "extra": {
         "symfony": {
-            "require": "5.4.*",
+            "require": "6.4.*",
             "allow-contrib": true
         },
         "branch-alias": {
-            "dev-master": "0.3.x-dev"
+            "dev-master": "0.4.x-dev"
         }
     },
     "config": {
