@@ -4,20 +4,17 @@ namespace Dontdrinkandroot\GitkiBundle\Analyzer;
 
 use Dontdrinkandroot\GitkiBundle\Model\Document\AnalyzedDocument;
 use Dontdrinkandroot\Path\FilePath;
+use Override;
 
 class TextAnalyzer implements AnalyzerInterface
 {
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function supports(FilePath $filePath, ?string $mimeType): bool
     {
         return $mimeType === 'text/plain';
     }
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function analyze(FilePath $path, $content): AnalyzedDocument
     {
         $analyzedFile = new AnalyzedDocument(
