@@ -5,15 +5,16 @@ namespace Dontdrinkandroot\GitkiBundle\Command;
 use Dontdrinkandroot\GitkiBundle\Service\Elasticsearch\ElasticsearchServiceInterface;
 use Dontdrinkandroot\GitkiBundle\Service\Elasticsearch\NoopElasticsearchService;
 use Dontdrinkandroot\GitkiBundle\Service\Wiki\WikiService;
+use Override;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
+#[AsCommand(name: 'gitki:reindex', description: 'Reindex all documents')]
 class ReindexCommand extends Command
 {
-    protected static $defaultDescription = 'Reindex all Markdown documents';
-
     public function __construct(
         private readonly WikiService $wikiService,
         private readonly ElasticsearchServiceInterface $elasticsearchService
@@ -21,18 +22,7 @@ class ReindexCommand extends Command
         parent::__construct();
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function configure(): void
-    {
-        $this
-            ->setName('gitki:reindex');
-    }
-
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         if ($this->elasticsearchService instanceof NoopElasticsearchService) {
