@@ -2,6 +2,7 @@
 
 namespace Dontdrinkandroot\GitkiBundle\Form\Type;
 
+use Override;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
@@ -9,9 +10,7 @@ use Symfony\Component\Form\FormBuilderInterface;
 
 class SubdirectoryCreateType extends AbstractType
 {
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
