@@ -7,7 +7,7 @@ namespace Dontdrinkandroot\GitkiBundle\Security;
  */
 class SecurityAttribute
 {
-    final public const READ_HISTORY = 'GITKI_READ_HISTORY';
-    final public const READ_PATH = 'GITKI_READ_PATH';
-    final public const WRITE_PATH = 'GITKI_WRITE_PATH';
+    final public const string READ_HISTORY = 'GITKI_READ_HISTORY';
+    final public const string READ_PATH = 'GITKI_READ_PATH';
+    final public const string WRITE_PATH = 'GITKI_WRITE_PATH';
 }
