@@ -2,27 +2,27 @@
 
 namespace Dontdrinkandroot\GitkiBundle\Tests\Acceptance;
 
+use Dontdrinkandroot\Common\Asserted;
 use Dontdrinkandroot\GitkiBundle\Tests\GitRepositoryTestTrait;
+use Dontdrinkandroot\GitkiBundle\Tests\TestApp\Security\StaticUserProvider;
+use Dontdrinkandroot\GitkiBundle\Tests\TestApp\Security\User;
+use Override;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase as BaseWebTestCase;
 
 abstract class WebTestCase extends BaseWebTestCase
 {
     use GitRepositoryTestTrait;
 
-    public const GIT_REPOSITORY_PATH = '/tmp/gitkitest/repo/';
+    public const string GIT_REPOSITORY_PATH = '/tmp/gitkitest/repo/';
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function setUp(): void
     {
         parent::setUp();
         $this->setUpRepo();
     }
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function tearDown(): void
     {
         parent::tearDown();
@@ -30,15 +30,32 @@ abstract class WebTestCase extends BaseWebTestCase
     }
 
     /**
-     * {@inheritdoc}
-     *
-     * @return string
-     *
      * @psalm-return '/tmp/gitkitest/repo/'
      */
-    protected function getRepositoryTargetPath()
+    #[Override]
+    protected function getRepositoryTargetPath(): string
     {
         return self::GIT_REPOSITORY_PATH;
+    }
+
+    /**
+     * @template T
+     * @param class-string<T> $class
+     * @return T
+     */
+    protected static function getService(string $class, ?string $id = null): object
+    {
+        if (null === $id) {
+            $id = $class;
+        }
+
+        return Asserted::instanceOf(self::getContainer()->get($id), $class);
+    }
+
+    protected static function getUser(string $identifier): User
+    {
+        $user = self::getService(StaticUserProvider::class)->loadUserByIdentifier($identifier);
+        return Asserted::instanceOf($user, User::class);
     }
 
     abstract protected function getEnvironment(): string;
