@@ -2,36 +2,31 @@
 
 namespace Dontdrinkandroot\GitkiBundle\Tests;
 
+use Override;
 use PHPUnit\Framework\TestCase;
 
 class GitRepositoryTestCase extends TestCase
 {
-    final const GIT_REPOSITORY_PATH = '/tmp/gitkitest/repo/';
+    final public const string GIT_REPOSITORY_PATH = '/tmp/gitkitest/repo/';
 
     use GitRepositoryTestTrait;
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function setUp(): void
     {
         $this->setUpRepo();
     }
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function tearDown(): void
     {
         $this->tearDownRepo();
     }
 
     /**
-     * {@inheritdoc}
-     *
-     *
      * @psalm-return '/tmp/gitkitest/repo/'
      */
+    #[Override]
     protected function getRepositoryTargetPath(): string
     {
         return self::GIT_REPOSITORY_PATH;
