@@ -2,10 +2,14 @@
 
 namespace Dontdrinkandroot\GitkiBundle\Tests\TestApp\Security;
 
+use Override;
 use Symfony\Component\Security\Core\Exception\UserNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
 
+/**
+ * @implements UserProviderInterface<User>
+ */
 class StaticUserProvider implements UserProviderInterface
 {
     /** @var array<string, User> */
@@ -18,17 +22,7 @@ class StaticUserProvider implements UserProviderInterface
         $this->users['admin'] = new User('admin', 'Mary Dane', 'marydane@examle.com', ['ROLE_USER', 'ROLE_ADMIN']);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function loadUserByUsername(string $username): UserInterface
-    {
-        return $this->loadUserByIdentifier($username);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function loadUserByIdentifier(string $identifier): UserInterface
     {
         if (!array_key_exists($identifier, $this->users)) {
@@ -38,17 +32,13 @@ class StaticUserProvider implements UserProviderInterface
         return $this->users[$identifier];
     }
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function refreshUser(UserInterface $user): UserInterface
     {
-        return $user;
+        return $this->loadUserByIdentifier($user->getUserIdentifier());
     }
 
-    /**
-     * {@inheritdoc}
-     */
+    #[Override]
     public function supportsClass(string $class): bool
     {
         return $class === User::class;
